requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic==1.10.12